pm-order-optimizer-backend/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
click==8.1.7
fastapi==0.109.2
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httptools==0.6.1
idna==3.6
joblib==1.3.2
k-means-constrained==0.7.3
mysqlclient==2.2.4
numpy==1.26.4
ortools==9.8.3296
pandas==2.2.0
protobuf==4.25.2
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
scikit-learn==1.4.0
scipy==1.12.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.26
starlette==0.36.3
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.4
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
